backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# Core Libraries
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2

# Async Tasks
celery==5.3.4

# Monitoring & Logging
python-dotenv==1.0.0
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0

# Image Processing (lightweight)
Pillow==10.2.0
